// frontend/hooks/models/adminReviewItemDTOPriorityTier.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export type AdminReviewItemDTOPriorityTier = typeof AdminReviewItemDTOPriorityTier[keyof typeof AdminReviewItemDTOPriorityTier];


export const AdminReviewItemDTOPriorityTier = {
  CONVERGING_REPORTS: 'CONVERGING_REPORTS',
  REPORTED_NO_CONVERGENCE: 'REPORTED_NO_CONVERGENCE',
  CONFIRMED_UNREPORTED: 'CONFIRMED_UNREPORTED',
  UNCONFIRMED_UNREPORTED: 'UNCONFIRMED_UNREPORTED',
} as const;