// frontend/hooks/models/adminReviewItemDTOVerificationStatus.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export type AdminReviewItemDTOVerificationStatus = typeof AdminReviewItemDTOVerificationStatus[keyof typeof AdminReviewItemDTOVerificationStatus];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const AdminReviewItemDTOVerificationStatus = {
  UNVERIFIED: 'UNVERIFIED',
  VERIFIED: 'VERIFIED',
  NEEDS_REVIEW: 'NEEDS_REVIEW',
  MANUAL_ENTRY: 'MANUAL_ENTRY',
} as const;