// frontend/hooks/models/resolveReportRequestVerificationStatus.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export type ResolveReportRequestVerificationStatus = typeof ResolveReportRequestVerificationStatus[keyof typeof ResolveReportRequestVerificationStatus];


export const ResolveReportRequestVerificationStatus = {
  UNVERIFIED: 'UNVERIFIED',
  VERIFIED: 'VERIFIED',
  NEEDS_REVIEW: 'NEEDS_REVIEW',
  MANUAL_ENTRY: 'MANUAL_ENTRY',
} as const;